requirements.txt
bcrypt==2.0.0
cffi==1.1.2
cryptography==1.0.1
dateutils==0.6.6
decorator==4.0.4
enum34==1.0.4
facebook-sdk==0.4.0
facepy==1.0.7
feedgen==0.3.1
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Cors==2.0.1
Flask-Login==0.2.11
Flask-Session==0.2
Flask-SQLAlchemy==2.0
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
ndg-httpsclient==0.4.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
pprintpp==0.2.3
ptyprocess==0.5
pyasn1==0.1.8
pycparser==2.14
pymongo==3.0.3
PyMySQL==0.6.6
pyOpenSSL==0.15.1
python-bcrypt==0.3.1
python-dateutil==2.4.2
pytz==2015.6
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.5
traitlets==4.0.0
Werkzeug==0.10.4
wheel==0.24.0
